% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks_metric.R
\name{ks_metric}
\alias{ks_metric}
\title{Customize the KS metric for Xgboost training.}
\usage{
ks_metric(yhat, dtrain)
}
\arguments{
\item{yhat}{predictions.}

\item{dtrain}{the input of the function \code{xgb.train}.}
}
\description{
Customize the KS metric for Xgboost training.
}
