% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgb_sql_score_query.R
\name{xgb_sql_score_query}
\alias{xgb_sql_score_query}
\title{Generate the score query from xgb object}
\usage{
xgb_sql_score_query(list_of_trees, features_table, feature_dict,
  base_score = 0.5, key_field = "id")
}
\arguments{
\item{base_score}{Extract the scale value from \code{xgb_model$params$base_score}.}
}
\description{
The core element of the add2xgb package is an object of the function xgb_sql_score_query.
It contains the process to translate XGBoost tree structure to SQL statements.
}
