% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgb_tree_sql.R
\name{xgb_tree_sql}
\alias{xgb_tree_sql}
\title{Translate decision tree into case when statement.}
\usage{
xgb_tree_sql(tree, feature_dict, sig = 5)
}
\description{
Translate decision tree into case when statement.
}
